check_license script: move in folder
